move DefConflictAnalysisPass
